requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
pydantic==2.6.1
python-dotenv==1.0.1 